+++ check/Control.Monad.Compat.check.hs
@@ -1,0 +1,3 @@
+{-# LANGUAGE NoImplicitPrelude #-}
+module Test where
+import Control.Monad.Compat